+++ src/RoadRunnerHttpApplicationRunner.php
@@ -34,9 +34,9 @@ use Yiisoft\Yii\Runner\Http\HttpApplicationRunner;
 use function gc_collect_cycles;
 
 /**
- * `RoadRunnerApplicationRunner` runs the Yii HTTP application using RoadRunner.
+ * `RoadRunnerHttpApplicationRunner` runs the Yii HTTP application using RoadRunner.
  */
-final class RoadRunnerApplicationRunner extends ApplicationRunner
+final class RoadRunnerHttpApplicationRunner extends ApplicationRunner
 {
     private ?ErrorHandler $temporaryErrorHandler = null;
     private ?PSR7WorkerInterface $psr7Worker = null;
@@ -45,13 +45,53 @@ final class RoadRunnerApplicationRunner extends ApplicationRunner
     /**
      * @param string $rootPath The absolute path to the project root.
      * @param bool $debug Whether the debug mode is enabled.
+     * @param bool $checkEvents Whether to check events' configuration.
      * @param string|null $environment The environment name.
+     * @param string $bootstrapGroup The bootstrap configuration group name.
+     * @param string $eventsGroup The events' configuration group name.
+     * @param string $diGroup The container definitions' configuration group name.
+     * @param string $diProvidersGroup The container providers' configuration group name.
+     * @param string $diDelegatesGroup The container delegates' configuration group name.
+     * @param string $diTagsGroup The container tags' configuration group name.
+     * @param string $paramsGroup The configuration parameters group name.
+     * @param array $nestedParamsGroups Configuration group names that are included into configuration parameters group.
+     * This is needed for recursive merging of parameters.
+     * @param array $nestedEventsGroups Configuration group names that are included into events' configuration group.
+     * This is needed for reverse and recursive merge of events' configurations.
+     *
+     * @psalm-param list<string> $nestedParamsGroups
+     * @psalm-param list<string> $nestedEventsGroups
      */
-    public function __construct(string $rootPath, bool $debug, ?string $environment)
-    {
-        parent::__construct($rootPath, $debug, $environment);
-        $this->bootstrapGroup = 'bootstrap-web';
-        $this->eventsGroup = 'events-web';
+    public function __construct(
+        string $rootPath,
+        bool $debug = false,
+        bool $checkEvents = false,
+        ?string $environment = null,
+        string $bootstrapGroup = 'bootstrap-web',
+        string $eventsGroup = 'events-web',
+        string $diGroup = 'di-web',
+        string $diProvidersGroup = 'di-providers-web',
+        string $diDelegatesGroup = 'di-delegates-web',
+        string $diTagsGroup = 'di-tags-web',
+        string $paramsGroup = 'params-web',
+        array $nestedParamsGroups = ['params'],
+        array $nestedEventsGroups = ['events'],
+    ) {
+        parent::__construct(
+            $rootPath,
+            $debug,
+            $checkEvents,
+            $environment,
+            $bootstrapGroup,
+            $eventsGroup,
+            $diGroup,
+            $diProvidersGroup,
+            $diDelegatesGroup,
+            $diTagsGroup,
+            $paramsGroup,
+            $nestedParamsGroups,
+            $nestedEventsGroups,
+        );
     }
 
     /**
@@ -106,16 +146,15 @@ final class RoadRunnerApplicationRunner extends ApplicationRunner
         $temporaryErrorHandler = $this->createTemporaryErrorHandler();
         $this->registerErrorHandler($temporaryErrorHandler);
 
-        $config = $this->getConfig();
-        $container = $this->getContainer($config, 'web');
+        $container = $this->getContainer();
 
         // Register error handler with real container-configured dependencies.
         /** @var ErrorHandler $actualErrorHandler */
         $actualErrorHandler = $container->get(ErrorHandler::class);
         $this->registerErrorHandler($actualErrorHandler, $temporaryErrorHandler);
 
-        $this->runBootstrap($config, $container);
-        $this->checkEvents($config, $container);
+        $this->runBootstrap();
+        $this->checkEvents();
 
         $env = Environment::fromGlobals();